refactor(evaluations): CreateQuestions sous 300 lignes — éclatement complet (H1)
CreateQuestions.vue 596 → 110 lignes. Éclatement façon G1 :
- useCreateQuestions.js (275) : chargement KLASSCI (+ repli dashboard) et LMS
  (mode édition), gestion questions/options/réponses, sauvegarde création
  (+publication) ou mise à jour.
- QuestionsConfigForm.vue (62) : section « Configuration ».
- QuestionItemEditor.vue (192) : carte question riche (surlignage, icônes A/B/C).
- QuestionsEditorList.vue (51) : liste + ajout + état vide, relais avec index.

CSS `.dashboard-content` conservé verbatim dans l'orchestrateur. Comportement,
services, routes, logs et alertes identiques. Tous fichiers ≤300.

Tests : 15 verts (composable création+édition + 3 montages). Build vert.
